+++ src/main/java/javaProjects/HRMS/business/abstracts/Users/CandidateService.java
@@ -3,11 +3,11 @@ import javaProjects.HRMS.core.business.abstracts.BaseService;
 import javaProjects.HRMS.core.utilities.results.DataResult;
 import javaProjects.HRMS.entities.concretes.Users.Candidate;
 
-public interface CandidateService extends BaseService<Candidate,Integer> {
+public interface CandidateService extends BaseService<Candidate,Long> {
 
 	DataResult<Candidate> getByIdentityNumber(String identityNumber);
 	DataResult<Candidate> getByEmail(String email);
-	DataResult<Candidate> getById(int id);
+	DataResult<Candidate> getById(Long id);
 	void save(Candidate candidate);
-	DataResult<Integer> add(Candidate candidate);
+	DataResult<Long> add(Candidate candidate);
 }
