+++ src/main/java/javaProjects/HRMS/business/abstracts/Users/SystemEmployeeService.java
@@ -4,6 +4,6 @@ import javaProjects.HRMS.core.business.abstracts.BaseService;
 import javaProjects.HRMS.core.utilities.results.Result;
 import javaProjects.HRMS.entities.concretes.Users.SystemEmployee;
 
-public interface SystemEmployeeService  extends BaseService<SystemEmployee, Integer> {
+public interface SystemEmployeeService  extends BaseService<SystemEmployee, Long> {
 
 }
