+++ src/main/java/javaProjects/HRMS/dataAccess/abstracts/Users/EmployerDao.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import javaProjects.HRMS.entities.concretes.Users.Employer;
 
-public interface EmployerDao extends JpaRepository<Employer, Integer>{
+public interface EmployerDao extends JpaRepository<Employer, Long>{
 	Employer getByEmail(String email);
 }
